fix: centralize extractIssueNumberFromBranch in core
- Add extractIssueNumberFromBranch to core git-utils
- Export from core index
- Update CLI to import from core and re-export
- Update VS Code to import from core via git-utils

Relates to #65
