+++ packages/cli/src/branch-linker.ts
@@ -11,6 +11,7 @@ import {
     parseBranchLink,
     setBranchLinkInBody,
     removeBranchLinkFromBody,
+    extractIssueNumberFromBranch,
     type RepoInfo,
 } from '@bretwardjames/ghp-core';
 import { api } from './github-api.js';
@@ -75,31 +76,8 @@ export async function getBranchForIssue(
     }
 }
 
-/**
- * Extract issue number from a branch name.
- * Supports common patterns:
- * - user/123-feature-name
- * - feature/123-something
- * - 123-fix-bug
- * - fix-123-something
- */
-export function extractIssueNumberFromBranch(branchName: string): number | null {
-    const patterns = [
-        /\/(\d+)-/,      // user/123-title
-        /^(\d+)-/,       // 123-title
-        /-(\d+)-/,       // feature-123-title
-        /[/#](\d+)$/,    // ends with #123 or /123
-    ];
-
-    for (const pattern of patterns) {
-        const match = branchName.match(pattern);
-        if (match) {
-            return parseInt(match[1], 10);
-        }
-    }
-
-    return null;
-}
+// Re-export from core for backwards compatibility
+export { extractIssueNumberFromBranch };
 
 /**
  * Result of finding an issue for a branch.
